// constants/enums/style-enum.ts
export enum Size {
  XS = 'x-small',
  S = 'small',
  M = 'default',
  L = 'large',
  XL = 'x-large',
}

export enum Color {
  Primary = 'primary',
  Secondary = 'secondary',
  Success = 'success',
  Info = 'info',
  Warning = 'warning',
  Error = 'error',
}

export enum Variant {
  Flat = 'flat',
  Text = 'text',
  Elevated = 'elevated',
  Tonal = 'tonal',
  Outlined = 'outlined',
  Plain = 'plain',
}

export enum FormSize {
  Default = 'default',
  Comfortable = 'comfortable',
  Compact = 'compact',
}

export enum FormVariant {
  Regular = 'regular',
  Outlined = 'outlined',
  Plain = 'plain',
  Underlined = 'underlined',
  Filled = 'filled',
  Solo = 'solo',
  SoloInvert = 'solo-inverted',
  SoloFilled = 'solo-filled',
}

export enum Direction {
  Horizontal = 'horizontal',
  Vertical = 'vertical',
}

export enum CardVariant {
  Flat = 'flat',
  Text = 'text',
  Elevated = 'elevated',
  Tonal = 'tonal',
  Outlined = 'outlined',
  Plain = 'plain',
}